cortex_code/requirements.txt
snowflake-snowpark-python==1.14.0
snowflake-connector-python==3.8.0
